+++ commandset/Services/GetCurrentViewInfoEventHandler.cs
@@ -43,7 +43,7 @@ namespace RevitMCPCommandSet.Services
                     DetailLevel = activeView.DetailLevel.ToString(),
                 };
             }
-            catch (Exception ex)
+            catch (Exception)
             {
                 TaskDialog.Show("error", "获取信息失败");
             }